Planos e agenda
Hardcoded todos os planos no planoseeder
Alterado agendacard
Alterado Planos.vue para construir tabela dos planos

+++ database/migrations/2022_07_30_144611_create_planos_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('tempo');
             $table->string('frequencia');
             $table->string('ferias')->nullable();
-            $table->decimal('valorTotal',5,2);
+            $table->float('valorTotal');
             $table->timestamps();
         });
     }
